+++ roqoqo/tests/integration/main.rs
@@ -21,3 +21,6 @@ mod circuit;
 
 #[cfg(test)]
 mod devices;
+
+#[cfg(test)]
+mod quantum_program;